Add files via upload
The purpose of this exercise is to take a look at the variables in the socioeconomic indicators dataset and to show the following skills:

 1. ETL(Extract, Transform, Load) data from an external source and upload it to a SQL database
 2. Perform SQL Queries to perform exploratory analysis of data
 3. Create graphs/visualizations from data to discover insights and patterns from the data set

Data Source: The city of Chicago released a dataset of socioeconomic data to the Chicago City Portal. This data set contains a selection of six socioeconomic indicators of public health significance and a “hardship index,” for each Chicago community area, for the years 2008 – 2012. Scores on the hardship index can range from 1 to 100, with a higher index number representing a greater level of hardship.

A detailed description of the data set can be found on [the city of Chicago's website](https://data.cityofchicago.org/Health-Human-Services/Census-Data-Select)