Add files via upload
We use this classification algorithm to build a model from the historical data of patients, and their response to different medications. Then we will be able in the future to use the trained decision tree to predict the class of an unknown patient, or to find a proper drug for a new patient.