Add files via upload
We apply a K-means cluster algorithm for customer segmentation. Customer segmentation is the practice of partitioning a customer base into groups of individuals that have similar characteristics. It is a significant strategy as a business can target these specific groups of customers and effectively allocate marketing resources. For example, one group may contain customers who are high-profit and low-risk, that is, more likely to purchase products, or subscribe for a service. A goal for a company could be to retain those customers.